Move RSA to public headers
