+++ src/apps/js_generic/js_generic_base.cpp
@@ -2,12 +2,12 @@
 // Licensed under the Apache 2.0 License.
 #include "apps/utils/metrics_tracker.h"
 #include "ccf/app_interface.h"
+#include "ccf/crypto/rsa_key_pair.h"
 #include "ccf/historical_queries_adapter.h"
 #include "ccf/user_frontend.h"
 #include "ccf/version.h"
 #include "crypto/entropy.h"
 #include "crypto/key_wrap.h"
-#include "crypto/rsa_key_pair.h"
 #include "js/wrap.h"
 #include "kv/untyped_map.h"
 #include "named_auth_policies.h"
