+++ src/crypto/key_wrap.cpp
@@ -4,9 +4,9 @@
 #include "key_wrap.h"
 
 #include "ccf/crypto/key_pair.h"
+#include "ccf/crypto/rsa_key_pair.h"
 #include "crypto/symmetric_key.h"
 #include "openssl/symmetric_key.h"
-#include "rsa_key_pair.h"
 
 #include <cstdint>
 #include <stdexcept>
