+++ src/crypto/key_wrap.h
@@ -2,8 +2,8 @@
 // Licensed under the Apache 2.0 License.
 #pragma once
 
+#include "ccf/crypto/rsa_key_pair.h"
 #include "crypto/symmetric_key.h"
-#include "rsa_key_pair.h"
 
 #include <cstddef>
 #include <cstdint>
