+++ src/crypto/openssl/rsa_public_key.cpp
@@ -2,8 +2,8 @@
 // Licensed under the Apache 2.0 License.
 
 #include "crypto/openssl/hash.h"
+#include "crypto/openssl/rsa_key_pair.h"
 #include "openssl_wrappers.h"
-#include "rsa_key_pair.h"
 
 namespace crypto
 {
