+++ src/crypto/openssl/rsa_public_key.h
@@ -3,9 +3,9 @@
 #pragma once
 
 #include "ccf/crypto/key_pair.h"
+#include "ccf/crypto/rsa_public_key.h"
 #include "crypto/openssl/hash.h"
 #include "crypto/openssl/public_key.h"
-#include "crypto/rsa_public_key.h"
 
 #include <optional>
 #include <string>
