+++ src/crypto/openssl/verifier.cpp
@@ -6,7 +6,7 @@
 #include "ccf/crypto/public_key.h"
 #include "ccf/ds/logger.h"
 #include "crypto/openssl/openssl_wrappers.h"
-#include "rsa_key_pair.h"
+#include "crypto/openssl/rsa_key_pair.h"
 #include "x509_time.h"
 
 #include <openssl/evp.h>
