+++ src/crypto/symmetric_key.cpp
@@ -3,7 +3,7 @@
 
 #include "crypto/openssl/symmetric_key.h"
 
-#include "crypto/rsa_key_pair.h"
+#include "ccf/crypto/rsa_key_pair.h"
 #include "symmetric_key.h"
 
 namespace crypto
