+++ src/node/share_manager.h
@@ -2,9 +2,9 @@
 // Licensed under the Apache 2.0 License.
 #pragma once
 
+#include "ccf/crypto/rsa_key_pair.h"
 #include "ccf/ds/logger.h"
 #include "crypto/entropy.h"
-#include "crypto/rsa_key_pair.h"
 #include "crypto/symmetric_key.h"
 #include "kv/encryptor.h"
 #include "ledger_secrets.h"
